move new query runner tests into submodule
